refactor(student) minor change to student hover - Change hover to be scale down rather than scale up which was causing issues in surroundings - Moved students to index.tsx rather than have a file in the pages folder to serve students
